+++ demo/src/Components/BottomSheet/BottomSheet.tsx
@@ -23,9 +23,9 @@ import type { HvProps, HvStyles } from './types';
 import Hyperview, { Events } from 'hyperview';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { dragDownHelper, dragUpHelper } from './utils';
-import { BottomSheetContentSection } from './BottomSheetContentSection';
+import { BottomSheetContentSection } from './ContentSection';
 import { Context as BottomSheetContext } from '../../Contexts/BottomSheet';
-import { BottomSheetStopPoint } from './BottomSheetStopPoint';
+import { BottomSheetStopPoint } from './StopPoint';
 import Overlay from './Overlay';
 import { namespace } from './types';
 import styles from './styles';
